# 100% readonly fields and ReadOnlySpan