--- a/.editorconfig
+++ b/.editorconfig
@@ -1,0 +1,22 @@
+[*.cs]
+end_of_line = lf
+indent_style = space
+indent_size = 4
+
+# IDE0049: Use framework type
+dotnet_style_predefined_type_for_locals_parameters_members = true
+dotnet_style_predefined_type_for_member_access = true
+
+# IDE0008: Use explicit type
+csharp_style_var_when_type_is_apparent = true
+csharp_style_var_for_built_in_types = true
+csharp_style_var_elsewhere = true
+
+# IDE0009: Member access should be qualified.
+dotnet_style_qualification_for_method = false
+dotnet_style_qualification_for_event = false
+dotnet_style_qualification_for_field = false
+dotnet_style_qualification_for_property = false
+
+# IDE0058: Expression value is never used
+dotnet_diagnostic.IDE0058.severity = none