--- a/src/Hashids.net/ArrayExtensions.cs
+++ b/src/Hashids.net/ArrayExtensions.cs
@@ -55,20 +55,8 @@ namespace HashidsNet
             ArrayPool<T>.Shared.Return(array);
         }
 
-        public static IEnumerable<Match> GetMatches(this MatchCollection matches) // Needed because prior to .NET 5, MatchCollection only implements IEnumerable, not IEnumerable<T>, so `foreach` gets #nullable warnings.
-        {
 #if NETCOREAPP3_1_OR_GREATER
-            return matches;
-#else
-            for (int i = 0; i < matches.Count; i++)
-            {
-                yield return matches[i];
-            }
-#endif
-        }
-
-#if NETCOREAPP3_1_OR_GREATER
-        /// <remarks>This method exists because <see cref="ReadOnlySpan{T}"/> does not implement <see cref="IEnumerable{T}"/>.</remarks>
+        /// <remarks>This method exists because <see cref="ReadOnlySpan{T}"/> does not implement <see cref="IEnumerable{T}"/> and using <c>.AsEnumerable()</c> will cause boxing.</remarks>
         public static bool Any<T>(this ReadOnlySpan<T> span, Func<T,bool> predicate)
         {
             for(int i = 0; i < span.Length; i++)