--- a/src/Hashids.net/Hashids.Obsolete.cs
+++ b/src/Hashids.net/Hashids.Obsolete.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.ComponentModel;
 
 namespace HashidsNet
 {
@@ -11,6 +12,7 @@ namespace HashidsNet
         /// Encodes the provided numbers into a hash.
         /// </summary>
         [Obsolete("Use 'Encode' instead. The method was renamed to better explain what it actually does.")]
+		[EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string Encrypt(params int[] numbers)
         {
             return Encode(numbers);
@@ -20,6 +22,7 @@ namespace HashidsNet
         /// Encrypts the provided hex-string to a hash.
         /// </summary>
         [Obsolete("Use 'EncodeHex' instead. The method was renamed to better explain what it actually does.")]
+		[EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string EncryptHex(string hex)
         {
             return EncodeHex(hex);
@@ -29,6 +32,7 @@ namespace HashidsNet
         /// Decodes the provided hash into an array of numbers.
         /// </summary>
         [Obsolete("Use 'Decode' instead. Method was renamed to better explain what it actually does.")]
+		[EditorBrowsable(EditorBrowsableState.Never)]
         public virtual int[] Decrypt(string hash)
         {
             return Decode(hash);
@@ -38,6 +42,7 @@ namespace HashidsNet
         /// Decodes the provided hash to a hex-string.
         /// </summary>
         [Obsolete("Use 'DecodeHex' instead. The method was renamed to better explain what it actually does.")]
+		[EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string DecryptHex(string hash)
         {
             return DecodeHex(hash);