--- a/src/Hashids.net/Hashids.Obsolete.cs
+++ b/src/Hashids.net/Hashids.Obsolete.cs
@@ -3,16 +3,13 @@ using System.ComponentModel;
 
 namespace HashidsNet
 {
-    /// <summary>
-    /// Generate YouTube-like hashes from one or many numbers. Use hashids when you do not want to expose your database ids to the user.
-    /// </summary>
     public partial class Hashids : IHashids
     {
         /// <summary>
         /// Encodes the provided numbers into a hash.
         /// </summary>
         [Obsolete("Use 'Encode' instead. The method was renamed to better explain what it actually does.")]
-		[EditorBrowsable(EditorBrowsableState.Never)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string Encrypt(params int[] numbers)
         {
             return Encode(numbers);
@@ -22,7 +19,7 @@ namespace HashidsNet
         /// Encrypts the provided hex-string to a hash.
         /// </summary>
         [Obsolete("Use 'EncodeHex' instead. The method was renamed to better explain what it actually does.")]
-		[EditorBrowsable(EditorBrowsableState.Never)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string EncryptHex(string hex)
         {
             return EncodeHex(hex);
@@ -32,7 +29,7 @@ namespace HashidsNet
         /// Decodes the provided hash into an array of numbers.
         /// </summary>
         [Obsolete("Use 'Decode' instead. Method was renamed to better explain what it actually does.")]
-		[EditorBrowsable(EditorBrowsableState.Never)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual int[] Decrypt(string hash)
         {
             return Decode(hash);
@@ -42,7 +39,7 @@ namespace HashidsNet
         /// Decodes the provided hash to a hex-string.
         /// </summary>
         [Obsolete("Use 'DecodeHex' instead. The method was renamed to better explain what it actually does.")]
-		[EditorBrowsable(EditorBrowsableState.Never)]
+        [EditorBrowsable(EditorBrowsableState.Never)]
         public virtual string DecryptHex(string hash)
         {
             return DecodeHex(hash);