--- a/test/Hashids.net.benchmark/Hashids.net.benchmark.csproj
+++ b/test/Hashids.net.benchmark/Hashids.net.benchmark.csproj
@@ -2,7 +2,7 @@
 
     <PropertyGroup>
         <OutputType>Exe</OutputType>
-        <TargetFramework>net5.0</TargetFramework>
+        <TargetFrameworks>net48;net5.0</TargetFrameworks>
         <IsPackable>false</IsPackable>
         <LangVersion>9</LangVersion>
     </PropertyGroup>