… src/Hashids.net/ReadOnlySpan.cs