@manigandham 2021-11-28
@Jehoel 

I only see 3 TODOs; - One about short variable names and the others about the seps/alphabet characters. We can take care of that after your other changes before doing a release.

This project was ported directly from the Javascript version originally so it followed the same layout and naming. Feel free to create a new branch or just push to this one and we can make another PR.